ViewControllerのDispatchable準拠をMainQueueDispatcherとして分離し、ViewModelにDIするよう修正
